+++ next.config.mjs
@@ -1,0 +1,8 @@
+// @ts-check
+
+/** @satisfies {import('next').NextConfig} */
+const nextConfig = {
+    reactStrictMode: true,
+};
+
+export default nextConfig;
