+++ tsconfig.json
@@ -17,8 +17,13 @@
         "paths": {
             "@/*": ["./src/*"],
             "@/public/*": ["./public/*"]
-        }
+        },
+        "plugins": [
+            {
+                "name": "next"
+            }
+        ]
     },
-    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+    "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
     "exclude": ["node_modules"]
 }
